chore: moved all pages to /pages dir
